requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
statsmodels>=0.14.0
pyyaml>=6.0
jupyter>=1.0.0
tqdm>=4.65.0
nbformat>=5.9.0
